requirements.txt
Flask==2.3.3
line-bot-sdk==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
urllib3==2.0.5
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
blinker==1.6.2
greenlet==2.0.2
